Move functions to super class so they are visible from BC transactions.
