+++ dotnet/src/dotnetframework/GxClasses/Middleware/GXHttp.cs
@@ -1405,6 +1405,7 @@ namespace GeneXus.Http
 				sendSpaHeaders();
 			}
 		}
+#if !NETCORE
 
 		protected string GetEncryptedHash(string value, string key)
 		{
@@ -1457,7 +1458,7 @@ namespace GeneXus.Http
 		{
 			return Decrypt64(value, key, true);
 		}
-
+#endif
 		protected string DecryptAjaxCall(string encrypted)
 		{
 			this.validEncryptedParm = false;
