+++ dotnet/src/dotnetframework/GxClasses/Model/gxproc.cs
@@ -122,54 +122,6 @@ namespace GeneXus.Procedure
 		protected virtual void printHeaders(){}
 		protected virtual void printFooters(){}
 
-        protected string GetEncryptedHash(string value, string key)
-        {
-            return Encrypt64(GXUtil.GetHash(GeneXus.Web.Security.WebSecurityHelper.StripInvalidChars(value), Cryptography.Constants.SecurityHashAlgorithm), key);
-        }
-
-		protected string Encrypt64(string value, string key)
-		{
-			return Encrypt64(value, key, false);
-		}
-		private string Encrypt64(string value, string key, bool safeEncoding)
-		{
-			string sRet = string.Empty;
-			try
-			{
-				sRet = Crypto.Encrypt64(value, key, safeEncoding);
-			}
-			catch (InvalidKeyException)
-			{
-				GXLogging.Error(log, "440 Invalid encryption key");
-			}
-			return sRet;
-		}
-		protected string UriEncrypt64(string value, string key)
-		{
-			return Encrypt64(value, key, true);
-		}
-
-		protected string Decrypt64(string value, string key)
-		{
-			return Decrypt64(value, key, false);
-		}
-		private string Decrypt64(string value, string key, bool safeEncoding)
-		{
-			String sRet = string.Empty;
-			try
-			{
-				sRet = Crypto.Decrypt64(value, key, safeEncoding);
-			}
-			catch (InvalidKeyException)
-			{
-				GXLogging.Error(log, "440 Invalid encryption key");
-			}
-			return sRet;
-		}
-		protected string UriDecrypt64(string value, string key)
-		{
-			return Decrypt64(value, key, true);
-		}
 		public msglist GX_msglist 
 		{
 			get	{ return context.GX_msglist ; }
